requirements.txt
faiss-cpu==1.7.2
fastapi==0.81.0
kornia==0.6.6
kornia_moons==0.2.1
lmdb==1.3.0
numba==0.56.0
numpy==1.23.2
opencv_python==4.6.0.66
Pillow==9.2.0
psycopg2==2.9.3
pydantic==1.9.2
tqdm==4.64.0
uvicorn==0.18.3
python-multipart
lru-dict
